Move form builder functions to separate file and add required fields to inputs and textareas
